formatting a lot of formatting, also added annotations
