+++ src/main/java/me/mooy1/infinityexpansion/Categories.java
@@ -8,19 +8,19 @@ import org.bukkit.NamespacedKey;
 public final class Categories {
 
     public static final Category INFINITY_MACHINES = new Category(new NamespacedKey(InfinityExpansion.getInstance(),
-            "INFINITY_MACHINES"),
-            new CustomItem(Material.SMITHING_TABLE, "&b&lInfinity &7&lMachines"),
-            2
+        "INFINITY_MACHINES"),
+        new CustomItem(Material.SMITHING_TABLE, "&b&lInfinity &7&lMachines"),
+        2
     );
     public static final Category INFINITY_MATERIALS = new Category(new NamespacedKey(InfinityExpansion.getInstance(),
-            "INFINITY_MATERIALS"),
-            new CustomItem(Material.NETHERITE_BLOCK, "&b&lInfinity &a&lMaterials"),
-            2
+        "INFINITY_MATERIALS"),
+        new CustomItem(Material.NETHERITE_BLOCK, "&b&lInfinity &a&lMaterials"),
+        2
     );
     public static final Category INFINITY_GEAR = new Category(new NamespacedKey(InfinityExpansion.getInstance(),
-            "INFINITY_GEAR"),
-            new CustomItem(Material.NETHERITE_CHESTPLATE, "&b&lInfinity &c&lGear"),
-            2
+        "INFINITY_GEAR"),
+        new CustomItem(Material.NETHERITE_CHESTPLATE, "&b&lInfinity &c&lGear"),
+        2
     );
 
     private Categories() { }
