+++ src/main/java/me/mooy1/infinityexpansion/machines/InfinityPanel.java
@@ -1,14 +1,14 @@
-package me.mooy1.infinityexpansion.Machines;
+package me.mooy1.infinityexpansion.machines;
 
-import io.github.thebusybiscuit.slimefun4.utils.ChestMenuUtils;
-import me.mooy1.infinityexpansion.Categories;
-import me.mooy1.infinityexpansion.Items;
 import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetProvider;
 import io.github.thebusybiscuit.slimefun4.core.networks.energy.EnergyNetComponentType;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunItems;
+import io.github.thebusybiscuit.slimefun4.utils.ChestMenuUtils;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import me.mooy1.infinityexpansion.Categories;
+import me.mooy1.infinityexpansion.Items;
 import me.mrCookieSlime.CSCoreLibPlugin.Configuration.Config;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
@@ -39,15 +39,15 @@ public class InfinityPanel extends SlimefunItem implements EnergyNetProvider, In
 
     public void setupInv() {
         createPreset(this, panel.getItem().getImmutableMeta().getDisplayName().orElse("&7THIS IS A BUG"),
-                blockMenuPreset -> {
-                    for (int i = 0; i < 9; i++)
-                        blockMenuPreset.addItem(i, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
+            blockMenuPreset -> {
+                for (int i = 0; i < 9; i++)
+                    blockMenuPreset.addItem(i, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
 
-                    blockMenuPreset.addItem(4,
-                            new CustomItem(Material.RED_STAINED_GLASS_PANE,
-                            "&cNot Generating..."),
-                            ChestMenuUtils.getEmptyClickHandler());
-                });
+                blockMenuPreset.addItem(4,
+                    new CustomItem(Material.RED_STAINED_GLASS_PANE,
+                        "&cNot Generating..."),
+                    ChestMenuUtils.getEmptyClickHandler());
+            });
     }
 
     public int getGeneratedOutput(@Nonnull Location l, @Nonnull Config data) {
@@ -72,22 +72,22 @@ public class InfinityPanel extends SlimefunItem implements EnergyNetProvider, In
 
         if (inv.toInventory() != null && !inv.toInventory().getViewers().isEmpty()) {
             inv.replaceExistingItem(4,
-                    canGenerate ? new CustomItem(Material.GREEN_STAINED_GLASS_PANE, "&aGeneration",
-                            "", "&bRate: " + generationType,
-                            "&7Generating at &6" + rate + " J/s ",
-                            "",
-                            "&7Stored: &6" + (stored + rate) + " J"
-                    )
-                            : new CustomItem(Material.ORANGE_STAINED_GLASS_PANE, "&cNot Generating",
-                            "", "&7Generator has reached maximum capacity.",
-                            "", "&7Stored: &6" + stored + " J")
+                canGenerate ? new CustomItem(Material.GREEN_STAINED_GLASS_PANE, "&aGeneration",
+                    "", "&bRate: " + generationType,
+                    "&7Generating at &6" + rate + " J/s ",
+                    "",
+                    "&7Stored: &6" + (stored + rate) + " J"
+                )
+                    : new CustomItem(Material.ORANGE_STAINED_GLASS_PANE, "&cNot Generating",
+                    "", "&7Generator has reached maximum capacity.",
+                    "", "&7Stored: &6" + stored + " J")
             );
         }
 
         return rate;
     }
 
-    public int getGeneratingAmount(@Nonnull Block block,  @Nonnull World world) {
+    public int getGeneratingAmount(@Nonnull Block block, @Nonnull World world) {
 
         if (world.getEnvironment() == World.Environment.NETHER) return this.panel.getDayGenerationRate();
         if (world.getEnvironment() == World.Environment.THE_END) return this.panel.getNightGenerationRate();
@@ -125,19 +125,20 @@ public class InfinityPanel extends SlimefunItem implements EnergyNetProvider, In
     public enum Panel {
 
         CELESTIAL(Items.CELESTIAL_PANEL, 1_800, 0, 300_000, RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
-                Items.MAGSTEEL, Items.MAGSTEEL, Items.MAGSTEEL,
-                Items.MACHINE_PLATE, SlimefunItems.SOLAR_GENERATOR_4, Items.MACHINE_PLATE,
-                Items.MACHINE_CIRCUIT, Items.MACHINE_CORE, Items.MACHINE_CIRCUIT
+            Items.MAGSTEEL, Items.MAGSTEEL, Items.MAGSTEEL,
+            Items.MACHINE_PLATE, SlimefunItems.SOLAR_GENERATOR_4, Items.MACHINE_PLATE,
+            Items.MACHINE_CIRCUIT, Items.MACHINE_CORE, Items.MACHINE_CIRCUIT
         }),
         VOID(Items.VOID_PANEL, 0, 5_400, 900_000, RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
-                Items.MAGNONIUM_INGOT, Items.MAGNONIUM_INGOT, Items.MAGNONIUM_INGOT,
-                Items.MACHINE_PLATE, Items.CELESTIAL_PANEL, Items.MACHINE_PLATE,
-                Items.MACHINE_CIRCUIT, Items.MACHINE_CORE, Items.MACHINE_CIRCUIT
+            Items.MAGNONIUM_INGOT, Items.MAGNONIUM_INGOT, Items.MAGNONIUM_INGOT,
+            Items.MACHINE_PLATE, Items.CELESTIAL_PANEL, Items.MACHINE_PLATE,
+            Items.MACHINE_CIRCUIT, Items.MACHINE_CORE, Items.MACHINE_CIRCUIT
         }),
-        INFINITY(Items.INFINITY_PANEL, 120_000, 120_000, 20_000_000, RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
-                Items.CELESTIAL_PANEL, Items.VOID_PANEL, Items.CELESTIAL_PANEL,
-                Items.INFINITY_INGOT, Items.INFINITE_CAPACITOR, Items.INFINITY_INGOT,
-                Items.INFINITE_MACHINE_CIRCUIT, Items.INFINITE_MACHINE_CORE, Items.INFINITE_MACHINE_CIRCUIT,
+        INFINITY(Items.INFINITY_PANEL, 120_000, 120_000, 20_000_000, RecipeType.ENHANCED_CRAFTING_TABLE,
+            new ItemStack[] {
+            Items.CELESTIAL_PANEL, Items.VOID_PANEL, Items.CELESTIAL_PANEL,
+            Items.INFINITY_INGOT, Items.INFINITE_CAPACITOR, Items.INFINITY_INGOT,
+            Items.INFINITE_MACHINE_CIRCUIT, Items.INFINITE_MACHINE_CORE, Items.INFINITE_MACHINE_CIRCUIT,
         });
 
         @Nonnull
