+++ src/main/java/me/mooy1/infinityexpansion/machines/Quarry.java
@@ -1,4 +1,4 @@
-package me.mooy1.infinityexpansion.Machines;
+package me.mooy1.infinityexpansion.machines;
 
 import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetComponent;
 import io.github.thebusybiscuit.slimefun4.core.networks.energy.EnergyNetComponentType;
@@ -23,9 +23,6 @@ import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.inventory.ItemStack;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
@@ -42,29 +39,30 @@ public class Quarry extends SlimefunItem implements InventoryBlock, EnergyNetCom
 
     private void setupInv() {
         createPreset(this, tier.getItem().getImmutableMeta().getDisplayName().orElse("&7THIS IS A BUG"),
-                blockMenuPreset -> {
-                    for (int i = 0; i < 13; i++) {
-                        blockMenuPreset.addItem(i, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
-                    }
-                    for (int i = 14; i < 28; i++) {
-                        blockMenuPreset.addItem(i, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
-                    }
-                    blockMenuPreset.addItem(35, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
-                    blockMenuPreset.addItem(36, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
-                    for (int i = 44; i < 54; i++) {
-                        blockMenuPreset.addItem(i, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
-                    }
-                    blockMenuPreset.addItem(13,
-                            new CustomItem(Material.RED_STAINED_GLASS_PANE,
-                                    "&cNot Mining..."),
-                            ChestMenuUtils.getEmptyClickHandler());
-                });
+            blockMenuPreset -> {
+                for (int i = 0; i < 13; i++) {
+                    blockMenuPreset.addItem(i, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
+                }
+                for (int i = 14; i < 28; i++) {
+                    blockMenuPreset.addItem(i, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
+                }
+                blockMenuPreset.addItem(35, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
+                blockMenuPreset.addItem(36, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
+                for (int i = 44; i < 54; i++) {
+                    blockMenuPreset.addItem(i, ChestMenuUtils.getBackground(), ChestMenuUtils.getEmptyClickHandler());
+                }
+                blockMenuPreset.addItem(13,
+                    new CustomItem(Material.RED_STAINED_GLASS_PANE,
+                        "&cNot Mining..."),
+                    ChestMenuUtils.getEmptyClickHandler());
+            });
     }
 
     @Override
     public void preRegister() {
         this.addItemHandler(new BlockTicker() {
             public void tick(Block b, SlimefunItem sf, Config data) { Quarry.this.tick(b); }
+
             public boolean isSynchronized() { return false; }
         });
     }
@@ -72,7 +70,7 @@ public class Quarry extends SlimefunItem implements InventoryBlock, EnergyNetCom
 
     @Override
     public int getCapacity() {
-        return tier.getEnergyConsumption()*2;
+        return tier.getEnergyConsumption() * 2;
     }
 
     @Nonnull
@@ -128,25 +126,25 @@ public class Quarry extends SlimefunItem implements InventoryBlock, EnergyNetCom
     @AllArgsConstructor(access = AccessLevel.PRIVATE)
     public enum Tier {
 
-        BASIC(Items.QUARRY, 1, 12, 2_400,RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
-                Items.MACHINE_PLATE, SlimefunItems.LARGE_CAPACITOR, Items.MACHINE_PLATE,
-                new ItemStack(Material.DIAMOND_PICKAXE), SlimefunItems.GEO_MINER, new ItemStack(Material.DIAMOND_PICKAXE),
-                Items.MACHINE_CIRCUIT , Items.MACHINE_CORE, Items.MACHINE_CIRCUIT
+        BASIC(Items.QUARRY, 1, 12, 2_400, RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
+            Items.MACHINE_PLATE, SlimefunItems.LARGE_CAPACITOR, Items.MACHINE_PLATE,
+            new ItemStack(Material.DIAMOND_PICKAXE), SlimefunItems.GEO_MINER, new ItemStack(Material.DIAMOND_PICKAXE),
+            Items.MACHINE_CIRCUIT, Items.MACHINE_CORE, Items.MACHINE_CIRCUIT
         }),
-        ADVANCED(Items.ADVANCED_QUARRY, 2, 8, 7_200,RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
-                Items.MACHINE_PLATE, SlimefunItems.CARBONADO_EDGED_CAPACITOR, Items.MACHINE_PLATE,
-                new ItemStack(Material.NETHERITE_PICKAXE), Items.QUARRY, new ItemStack(Material.NETHERITE_PICKAXE),
-                Items.MACHINE_CIRCUIT , Items.MACHINE_CORE, Items.MACHINE_CIRCUIT
+        ADVANCED(Items.ADVANCED_QUARRY, 2, 8, 7_200, RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
+            Items.MACHINE_PLATE, SlimefunItems.CARBONADO_EDGED_CAPACITOR, Items.MACHINE_PLATE,
+            new ItemStack(Material.NETHERITE_PICKAXE), Items.QUARRY, new ItemStack(Material.NETHERITE_PICKAXE),
+            Items.MACHINE_CIRCUIT, Items.MACHINE_CORE, Items.MACHINE_CIRCUIT
         }),
-        VOID(Items.VOID_QUARRY, 5, 4, 27_000,RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
-                Items.MACHINE_PLATE, SlimefunItems.ENERGIZED_CAPACITOR, Items.MACHINE_PLATE,
+        VOID(Items.VOID_QUARRY, 5, 4, 27_000, RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
+            Items.MACHINE_PLATE, SlimefunItems.ENERGIZED_CAPACITOR, Items.MACHINE_PLATE,
             Items.MAGNONIUM_PICKAXE, Items.ADVANCED_QUARRY, Items.MAGNONIUM_PICKAXE,
-                Items.MACHINE_CIRCUIT , Items.INFINITE_MACHINE_CORE, Items.MACHINE_CIRCUIT
+            Items.MACHINE_CIRCUIT, Items.INFINITE_MACHINE_CORE, Items.MACHINE_CIRCUIT
         }),
         INFINITY(Items.INFINITY_QUARRY, 25, 1, 240_000, RecipeType.ENHANCED_CRAFTING_TABLE, new ItemStack[] {
-                Items.INFINITY_INGOT, Items.INFINITE_CAPACITOR, Items.INFINITY_INGOT,
-                Items.INFINITY_PICKAXE, Items.VOID_QUARRY, Items.INFINITY_PICKAXE,
-                Items.INFINITE_MACHINE_CIRCUIT , Items.INFINITE_MACHINE_CORE, Items.INFINITE_MACHINE_CIRCUIT
+            Items.INFINITY_INGOT, Items.INFINITE_CAPACITOR, Items.INFINITY_INGOT,
+            Items.INFINITY_PICKAXE, Items.VOID_QUARRY, Items.INFINITY_PICKAXE,
+            Items.INFINITE_MACHINE_CIRCUIT, Items.INFINITE_MACHINE_CORE, Items.INFINITE_MACHINE_CIRCUIT
         });
