+++ src/main/java/me/mooy1/infinityexpansion/materials/VoidDustResource.java
@@ -1,28 +1,32 @@
-package me.mooy1.infinityexpansion.Materials;
+package me.mooy1.infinityexpansion.materials;
 
+import io.github.thebusybiscuit.slimefun4.api.geo.GEOResource;
 import me.mooy1.infinityexpansion.InfinityExpansion;
 import me.mooy1.infinityexpansion.Items;
-import io.github.thebusybiscuit.slimefun4.api.geo.GEOResource;
 import org.bukkit.NamespacedKey;
 import org.bukkit.World;
 import org.bukkit.block.Biome;
 import org.bukkit.inventory.ItemStack;
+
+import javax.annotation.Nonnull;
 
 public class VoidDustResource implements GEOResource {
 
     private final NamespacedKey key = new NamespacedKey(InfinityExpansion.getInstance(), "VoidDust");
 
     @Override
-    public int getDefaultSupply(World.Environment environment, Biome biome) {
+    public int getDefaultSupply(@Nonnull World.Environment environment, @Nonnull Biome biome) {
         if (environment == World.Environment.THE_END) {
             return 1;
-        } if (biome == Biome.THE_VOID) {
+        }
+        if (biome == Biome.THE_VOID) {
             return 1;
         } else {
             return 0;
         }
     }
 
+    @Nonnull
     @Override
     public NamespacedKey getKey() {
         return key;
@@ -33,11 +37,13 @@ public class VoidDustResource implements GEOResource {
         return 1;
     }
 
+    @Nonnull
     @Override
     public String getName() {
         return "Void Dust";
     }
 
+    @Nonnull
     @Override
     public ItemStack getItem() {
         return Items.VOID_DUST.clone();
